+++ components/Layout.tsx
@@ -1,0 +1,17 @@
+import React from 'react'
+import Navbar from './Navbar'
+import Footer from './Footer'
+
+type Props = {
+  children: React.ReactNode
+}
+
+export default function Layout({ children }: Props) {
+  return (
+    <div className="min-h-screen bg-black text-white">
+      <Navbar />
+      <main>{children}</main>
+      <Footer />
+    </div>
+  )
+}